Refactor handler and service structure for improved organization and clarity
