+++ .gitignore
@@ -14,7 +14,6 @@
 # Binaries
 cmd/shortener/main
 cmd/shortener/shortener
-shortener
 
 # Dependency directories (remove the comment below to include it)
 vendor/
