+++ internal/app/app.go
@@ -5,17 +5,17 @@ import (
 	"yp-go-short-url-service/internal/config"
 	"yp-go-short-url-service/internal/config/db"
 	"yp-go-short-url-service/internal/handler"
-	shortenAPI "yp-go-short-url-service/internal/handler/api/shorten"
 	"yp-go-short-url-service/internal/handler/health"
-	urlExtractorHandler "yp-go-short-url-service/internal/handler/urlExtractor"
-	urlShortenerHandler "yp-go-short-url-service/internal/handler/urlShortener"
+	urlExtractorHandler "yp-go-short-url-service/internal/handler/urls/extractor"
+	shortenAPI "yp-go-short-url-service/internal/handler/urls/shortener/json"
+	urlShortenerHandler "yp-go-short-url-service/internal/handler/urls/shortener/text"
 	"yp-go-short-url-service/internal/middleware"
 	"yp-go-short-url-service/internal/middleware/gzip"
 	"yp-go-short-url-service/internal/repository"
 	"yp-go-short-url-service/internal/repository/postgres"
 	healthService "yp-go-short-url-service/internal/service/health"
-	urlExtractorService "yp-go-short-url-service/internal/service/urlExtractor"
-	urlShortenerService "yp-go-short-url-service/internal/service/urlShortener"
+	urlExtractorService "yp-go-short-url-service/internal/service/urls/extractor"
+	urlShortenerService "yp-go-short-url-service/internal/service/urls/shortener"
 
 	_ "yp-go-short-url-service/docs"
