+++ internal/handler/urls/extractor/handler.go
@@ -1,4 +1,4 @@
-package urlExtractor
+package extractor
 
 import (
 	"net/http"
