+++ internal/handler/urls/extractor/handler_test.go
@@ -1,4 +1,4 @@
-package urlExtractor
+package extractor
 
 import (
 	"context"
