+++ internal/handler/urls/shortener/json/handler.go
@@ -1,4 +1,4 @@
-package shorten
+package json
 
 import (
 	"encoding/json"
