+++ internal/handler/urls/shortener/json/handler_test.go
@@ -1,4 +1,4 @@
-package shorten
+package json
 
 import (
 	"context"
