+++ internal/handler/urls/shortener/text/handler.go
@@ -1,4 +1,4 @@
-package urlShortener
+package text
 
 import (
 	"fmt"
