+++ internal/handler/urls/shortener/text/handler_test.go
@@ -1,4 +1,4 @@
-package urlShortener
+package text
 
 import (
 	"context"
