+++ internal/service/urls/extractor/service.go
@@ -1,4 +1,4 @@
-package urlExtractor
+package extractor
 
 import (
 	"context"
