+++ internal/service/urls/shortener/shorten_url_base62.go
@@ -1,4 +1,4 @@
-package urlShortener
+package shortener
 
 import (
 	"crypto/md5"
