+++ internal/service/urls/shortener/shorten_url_base62_test.go
@@ -1,4 +1,4 @@
-package urlShortener
+package shortener
 
 import (
 	"math/big"
